+++ CHANGES.rst
@@ -12,6 +12,11 @@
 Changes
 =======
 
+Version v16.5.1 (released 2024-12-16)
+
+- pids: add manage permission to be able to manage DOIs
+- deposit: fix validation check when user needs a DOI and DOI is optional
+
 Version v16.5.0 (released 2024-12-16)
 
 - pids: add support for optional DOI